+++ src/test/java/de/codescape/jira/plugins/scrumpoker/ScrumPokerResourceBundleTest.java
@@ -8,6 +8,7 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.IOException;
+import java.net.URISyntaxException;
 import java.net.URL;
 import java.util.Arrays;
 import java.util.List;
@@ -16,6 +17,7 @@ import java.util.Scanner;
 import java.util.stream.Collectors;
 
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.*;
 
 public class ScrumPokerResourceBundleTest {
 
@@ -28,7 +30,22 @@ public class ScrumPokerResourceBundleTest {
     private static final String UPPER_CASE_LETTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
     private static final String NUMBERS = "0123456789";
     private static final String SPECIAL_CHARACTERS = ".,!?= \\'{}:-+\"()_";
-    private static final String ALLOWED_CHARACTERS = LOWER_CASE_LETTERS + UPPER_CASE_LETTERS + NUMBERS + SPECIAL_CHARACTERS;
+    private static final String ALLOWED_CHARACTERS =
+        LOWER_CASE_LETTERS + UPPER_CASE_LETTERS + NUMBERS + SPECIAL_CHARACTERS;
+
+    /* tests for files */
+
+    @Test
+    public void resourceBundleContainsExpectedTranslations() throws Exception {
+        List<String> resourceBundles = findAllResourceBundles();
+        assertThat(resourceBundles.size(), is(equalTo(SUPPORTED_BUNDLES.length)));
+
+        Arrays.stream(SUPPORTED_BUNDLES)
+            .map(bundle -> bundle.substring(bundle.indexOf("/") + 1))
+            .forEach(bundle -> assertThat(resourceBundles, hasItem(bundle)));
+    }
+
+    /* tests for file content */
 
     @Test
     public void resourceBundlesMustOnlyContainAllowedCharacters() {
@@ -57,6 +74,18 @@ public class ScrumPokerResourceBundleTest {
                 }
             }
         );
+    }
+
+    /* supporting methods */
+
+    private List<String> findAllResourceBundles() throws URISyntaxException {
+        URL directory = getClass().getClassLoader().getResource("i18n");
+        assertThat("Directory must be found", directory, is(notNullValue()));
+
+        String[] bundles = new File(directory.toURI()).list();
+        assertThat("Directory must contain files", bundles, is(notNullValue()));
+
+        return Arrays.asList(bundles);
     }
 
     private boolean isNotACommentLine(String line) {
@@ -93,9 +122,7 @@ public class ScrumPokerResourceBundleTest {
 
     private File getFile(String fileName) {
         URL resource = getClass().getClassLoader().getResource(fileName);
-        if (resource == null) {
-            throw new RuntimeException("File " + fileName + " not found.");
-        }
+        assertThat("File must be found", resource, is(notNullValue()));
         return new File(resource.getFile());
     }